Fixed bug when installing from private git repo
git clone /path/to/foo@host.com:repo-HASH /path/to/foo@host.com:repo-HASH-ref

This will interpret /path/to/foo as the username to send to host.com,
when it was supposed to be a filepath.  Easy fix: use file:///path

It wasn't obvious to me how to test this appopriately, so sorry... no
tests attached.